Move LoginScreen to screens/login. Add UserInfo type in AuthContext state.
