+++ src/contexts/AuthContext.tsx
@@ -5,26 +5,33 @@ export enum AuthActionType {
     Logout = 'logout'
 }
 
-type Action = { type: AuthActionType, username?: string }
+type UserInfo = {
+    username: string,
+    eid?: string,
+};
+type Action = {
+    type: AuthActionType,
+    user?: UserInfo
+}
 type Dispatch = (action: Action) => void
 type State = {
     isLoggedIn: boolean,
-    username: string,
+    user: UserInfo,
 }
 const initialState: State = {
     isLoggedIn: false,
-    username: null
+    user: null,
 };
 
 const AuthStateContext = React.createContext<State | undefined>(undefined);
 const AuthDispatchContext = React.createContext<Dispatch | undefined>(undefined);
 
-function reducer(state: State, action: Action) {
+function reducer(state: State, action: Action): State {
     switch (action.type) {
         case AuthActionType.Login:
-            return {...state, isLoggedIn: true, username: action.username};
+            return {...state, isLoggedIn: true, user: action.user};
         case AuthActionType.Logout:
-            return {...state, isLoggedIn: false, username: null};
+            return {...state, isLoggedIn: false, user: null};
     }
 }
