+++ src/styles/base.ts
@@ -1,4 +1,4 @@
-import { StyleSheet } from 'react-native';
+import {Platform, StatusBar, StyleSheet} from 'react-native';
 
 export default StyleSheet.create({
     container: {
@@ -6,6 +6,7 @@ export default StyleSheet.create({
         alignItems: 'center',
         justifyContent: 'center',
         flexDirection: 'column',
+        paddingTop: Platform.OS === 'android' ? StatusBar.currentHeight : 0,
     },
     text: {
         fontSize: 14,
